update theme, update header, tested site
