+++ src/components/CartHeader/CartHeader.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import CartLogo from "../SVG/CartLogo";
+import BurgerLogo from "../SVG/BurgerLogo";
+import back from "../../assets/icons/back.svg";
+
+function CartHeader() {
+  return <div className="secondary__navigation"></div>;
+}
+
+export default CartHeader;
